// src/types/Api.ts
interface IWeatherApi {
    coord: Array<{
      lon: number;
      lat: number;
    }>;
    weather: Array<{
      id: number;
      main: string;
      description: string;
      icon: string;
    }>;
    base: string;
    name: string;
    main: {
      temp: number;
      feels_like: number;
      temp_max: number;
      temp_min: number;
      pressure: number;
      humidity: number;
    };
    visibility: number;
    wind: {
      speed: number;
      deg: number;
    };
    clouds: {
      all: number;
    };
    dt: number;
    sys: {
      type: number;
      id: number;
      country: string;
      sunrise: number;
      sunset: number;
    };
    timezone: number;
    id: number;
    cod: number;
}

export { IWeatherApi };